+++ history.md
@@ -1,3 +1,13 @@
+**2020.09.14**
+
+- add: album cover from ID3-Tag or file named `cover.jpg` in folder with playing music file;
+- add: turn random on/off over notification.
+
+**2020.09.06**
+
+- rm: `drive_path` variable;
+- upd: logs.
+
 **2020.09.04**
 
 - add: .flac and .wav file formats support;